+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [x]
+- Improve user selection in assistants and deputies
 - Request new users from requests table
 - Request create user and accept time/user
 - Some changes in request info